pyproject.toml
[project]
name = "a2a-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=0.1.1",
    "strands-agents[ollama]>=0.3.0",
]
